Initial FIle Upload
This is my first ever commit to this project. The project is not nearly done yet, but enough progress has been made now to officially start sharing with the community what I've created so far. This also serves as a helpful and easy way to back up my work, and track my progress as I go. Hopefully more commits and updates will be soon to follow!